[aotinductor] Move cache_dir to utils.py
Summary: Some tests can utilize cache_dir()

[ghstack-poisoned]
